Refactor styles and upgrade Sass version
This commit updates several components and styles to improve the overall look and functionality of the project. Major changes include revising styling for MKLayout, MKDropdown and MKMenu components, deleting the stories.scss file and introducing a new MKLayoutToggle component. The Sass package was also upgraded to version 1.71.0, which will bring in new features and bug fixes.