Refactor MKButton component to use styled-components
Replaced traditional CSS styling with `styled-components` in MKButton component to allow dynamic styling based on properties. This change included removal of the `style.scss` file, and replacing it with a `style.ts` file, containing emotion-styled components. Note that this refactor is part of an overall effort to improve the readability, maintainability, and modularity of our codebase.